+++ NEWS.md
@@ -3,7 +3,7 @@
 ## v0.0.2 - 2026-01-28
 ### Fixed  
 * error in `calc_exp_mo2s()` when there were no cycles with R^2 under the threshold.  
-
+* added `correct_for_temperature()` and `temp_correct_data`.
 
 ## v0.0.1